backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
httpx==0.27.2
boto3==1.35.0
python-dotenv==1.0.1
